requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx